Remove handles in Eternity
We don't need to evacuate immortal objects.  Without handles, can save
space in the Handle pool, plus maybe some time in GC.  On the downside,
we will create more handles when creating empty objects a lot, but it is
not sure if this really is a use case.  In addition, dedicated handles
per Thread might lessen contention in atomic reference counts.